backend/requirements.txt
asgiref==3.8.1
django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3
django-cacheops==7.2
redis==6.1.0
celery==5.5.2
django-celery-results==2.6.0
django-celery-beat==2.8.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-redis==5.4.0
